Joe/Joe_Bot_v1/requirements.txt
python-telegram-bot==20.7
anthropic>=0.25.0
gspread>=6.0.0
google-auth>=2.0.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pytz>=2024.1
